move a bunch of stuff around (#967)

+++ backend/pkg/types/fakes/workers.go
@@ -16,7 +16,7 @@ func BuildFakeFinalizeMealPlansRequest() *types.FinalizeMealPlansRequest {
 // BuildFakeFinalizeMealPlansResponse builds a faked FinalizeMealPlansResponse.
 func BuildFakeFinalizeMealPlansResponse() *types.FinalizeMealPlansResponse {
 	return &types.FinalizeMealPlansResponse{
-		Count: int(buildFakeNumber()),
+		Count: int64(buildFakeNumber()),
 	}
 }
